requirements.txt
appnope==0.1.0
backcall==0.2.0
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
decorator==4.4.2
html5lib==1.1
idna==2.10
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
kiwisolver==1.2.0
lxml==4.5.2
matplotlib==3.3.1
mojimoji==0.0.11
MouseInfo==0.1.3
mysqlclient==2.0.1
numpy==1.19.1
pandas==1.1.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.7
psycopg2==2.8.5
ptyprocess==0.6.0
PyAutoGUI==0.9.50
PyGetWindow==0.0.8
Pygments==2.6.1
PyMsgBox==1.0.8
pyparsing==2.4.7
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.1
PyTweening==1.0.3
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1